More style changes
Added upload/update dates to the search cards.
Swapped out the missing image texture for one that should be visible on both dark and light modes.
Fixed a bug where newly installed mods would not be loaded by the game unless you pressed the apply priority button. Now the manager checks to see if there is a loaded modpack and if there is not it will apply the priority after all the mods have been loaded in.