Config update, Remember all mods enabled states
the enabled state of the installed mods is now remembered for the "All Mods" default pack.
app will always check on launch if DML is installed, allowing users to reinstall dml if it's missing.

Latest commit hash is now displayed as part of the version number in debug builds.

Adding icons to most buttons missing an icon.